Redux of expo initial project to use App.js instead of typescript file
