--- .expo/README.md
@@ -1,17 +1,0 @@
-> Why do I have a folder named ".expo" in my project?
-
-The ".expo" folder is created when an Expo project is started using "expo start" command.
-
-> What does the "packager-info.json" file contain?
-
-The "packager-info.json" file contains port numbers and process PIDs that are used to serve the application to the mobile device/simulator.
-
-> What does the "settings.json" file contain?
-
-The "settings.json" file contains the server configuration that is used to serve the application manifest.
-
-> Should I commit the ".expo" folder?
-
-No, you should not share the ".expo" folder. It does not contain any information that is relevant for other developers working on the project, it is specific to your machine.
-
-Upon project creation, the ".expo" folder is already added to your ".gitignore" file.
