--- .expo/settings.json
@@ -1,10 +1,0 @@
-{
-  "hostType": "lan",
-  "lanType": "ip",
-  "dev": true,
-  "minify": false,
-  "urlRandomness": "py-uju",
-  "https": false,
-  "scheme": null,
-  "devClient": false
-}
