--- README.md
@@ -1,2 +1,0 @@
-# RockOn Mobile
-Frontend application for RockOn [Core](https://github.com/CIS3296SoftwareDesignF21/rockon-core)
