+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "myproject",
+  "name": "rockon-mobile",
   "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
@@ -7,38 +7,18 @@
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "eject": "expo eject",
-    "test": "jest --watchAll"
-  },
-  "jest": {
-    "preset": "jest-expo"
+    "eject": "expo eject"
   },
   "dependencies": {
-    "@expo/vector-icons": "^12.0.0",
-    "@react-navigation/bottom-tabs": "^6.0.5",
-    "@react-navigation/native": "^6.0.2",
-    "@react-navigation/native-stack": "^6.1.0",
-    "expo": "~43.0.0",
-    "expo-asset": "~8.4.3",
-    "expo-constants": "~12.1.2",
-    "expo-font": "~10.0.3",
-    "expo-linking": "~2.4.2",
-    "expo-splash-screen": "~0.13.3",
+    "expo": "~43.0.2",
     "expo-status-bar": "~1.1.0",
-    "expo-web-browser": "~10.0.3",
     "react": "17.0.1",
     "react-dom": "17.0.1",
-    "react-native": "0.64.2",
-    "react-native-safe-area-context": "3.3.2",
-    "react-native-screens": "~3.8.0",
+    "react-native": "0.64.3",
     "react-native-web": "0.17.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9",
-    "@types/react": "~17.0.21",
-    "@types/react-native": "~0.64.12",
-    "jest-expo": "~43.0.0",
-    "typescript": "~4.3.5"
+    "@babel/core": "^7.12.9"
   },
   "private": true
 }
